Migrated util to TS
