+++ hrm-service/src/async-handler.ts
@@ -1,0 +1,1 @@
+export default fn => (req, res, next) => fn(req, res, next).catch(next);
