+++ hrm-service/src/permissions/unauthorized.ts
@@ -1,10 +1,8 @@
-const asyncHandler = fn => (req, res, next) => fn(req, res, next).catch(next);
+import { Response } from 'express';
 
-const unauthorized = res => {
+export const unauthorized = (res: Response) => {
   const authorizationFailed = { error: 'Authorization failed' };
   console.log(`[authorizationMiddleware]: ${JSON.stringify(authorizationFailed)}`);
   res.status(401).json(authorizationFailed);
   return authorizationFailed;
 };
-
-module.exports = { asyncHandler, unauthorized };
