--- hrm-service/src/routes/index.ts
@@ -1,3 +1,0 @@
-import v0Routes from './v0';
-
-export const apiV0 = v0Routes;
